#1434 separating web socket api tests to logical test suits with their own table and module definitions
